test(EuiDataGrid): move EuiDataGrid row height options jest tests to a separate file
